refactor: split debug html
